pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.7.0"
description = ""
authors = ["Xodarap7 <basyar86@gmail.com>"]
repository = "https://github.com/Xodarap7/python-project-lvl3"
packages = [
  { include = "page_loader" },
]

[tool.poetry.dependencies]
python = "^3.9.12"
requests-mock = "^1.9.3"
progress = "^1.6"
beautifulsoup4 = "^4.11.1"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
pytest = "^7.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
page-loader = "page_loader.scripts.loader:main"
